move icon battery function from util to helpers
