+++ homeassistant/helpers/icon.py
@@ -1,4 +1,4 @@
-"""Icon util methods."""
+"""Icon helper methods."""
 from typing import Optional
 
 
@@ -11,8 +11,10 @@ def icon_for_battery_level(battery_level: Optional[int]=None,
     if charging and battery_level > 10:
         icon += '-charging-{}'.format(
             int(round(battery_level / 20 - .01)) * 20)
-    elif charging or battery_level <= 5:
+    elif charging:
         icon += '-outline'
+    elif battery_level <= 5:
+        icon += '-alert'
     elif 5 < battery_level < 95:
         icon += '-{}'.format(int(round(battery_level / 10 - .01)) * 10)
     return icon
